+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## GRNOC Worldview Plugin 1.1.0 -- Fri Sep 2021
+* More Universal Data Mapping
+* Bug Fixes
+
 ## GRNOC Worldview Plugin 1.0.1 -- Tue Apr 2021
 
 * Hides zoom in/zoom out/editing controls when in presentation view
 * Ability to have custom legend text colors
+
 ## GRNOC Worldview Plugin 1.0.0 -- Thu Apr 2021
 ### Key Features
 - Create network maps